fix: Pull request fixes on Product-mutations
