--- nerdery-ecommerce/src/common/inputs/toggle.input.ts
@@ -1,17 +1,0 @@
-import { InputType, Field, Int } from '@nestjs/graphql';
-import { IsInt, IsOptional, IsPositive } from 'class-validator';
-
-@InputType()
-export class ToggleInput {
-    @Field(() => Int, { nullable: true, defaultValue: 1 })
-    @IsInt()
-    @IsOptional()
-    @IsPositive()
-    page?: number;
-
-    @Field(() => Int, { nullable: true, defaultValue: 20 })
-    @IsInt()
-    @IsOptional()
-    @IsPositive()
-    limit?: number;
-}
